feat: refactor to separate CalendarDate
